Change parameters from 32bit to 64bit

+++ source/common/http/filter/fault_filter.h
@@ -38,8 +38,8 @@ struct FaultFilterHeaders {
  */
 struct FaultFilterConfig {
   FaultFilterConfig(const std::string& stat_prefix, Stats::Store& stats,
-                    Runtime::RandomGenerator& random, uint32_t abort_code,
-                    uint32_t abort_probability, uint32_t delay_probability,
+                    Runtime::RandomGenerator& random, uint64_t abort_code,
+                    uint64_t abort_probability, uint64_t delay_probability,
                     std::chrono::milliseconds delay_duration,
                     std::vector<FaultFilterHeaders> fault_filter_headers)
       : random_(random), modulo_base_(100), abort_code_(abort_code),
@@ -48,11 +48,11 @@ struct FaultFilterConfig {
         stats_{ALL_FAULT_FILTER_STATS(POOL_COUNTER_PREFIX(stats, stat_prefix))} {}
 
   Runtime::RandomGenerator& random_;
-  uint32_t modulo_base_; // % of traffic to inject faults on, in increments of 1%. Hardcoded to 100,
+  uint64_t modulo_base_; // % of traffic to inject faults on, in increments of 1%. Hardcoded to 100,
                          // until we need higher precision.
-  uint32_t abort_code_;  // HTTP or gRPC return codes
-  uint32_t abort_probability_;               // 0-100
-  uint32_t delay_probability_;               // 0-100
+  uint64_t abort_code_;  // HTTP or gRPC return codes
+  uint64_t abort_probability_;               // 0-100
+  uint64_t delay_probability_;               // 0-100
   std::chrono::milliseconds delay_duration_; // in milliseconds
   std::vector<FaultFilterHeaders> fault_filter_headers_;
   FaultFilterStats stats_;
